feat(tau-ops): split channel_store_admin below 2200 (#2044)
