+++ docs/README.md
@@ -23,6 +23,11 @@ This index maps Tau documentation by audience and task.
 | Runtime contributor | [Startup DI Pipeline](guides/startup-di-pipeline.md) | 3-stage startup resolution: preflight gate, dependency/context composition, mode dispatch |
 | Runtime contributor | [Contract Pattern Lifecycle](guides/contract-pattern-lifecycle.md) | Shared fixture lifecycle, compatibility gates, extension checklist, anti-patterns |
 | Runtime contributor | [Tool Name Registry](guides/tool-name-registry.md) | Reserved built-in tool-name catalog and registration conflict behavior for extension + MCP external tools |
+| Runtime contributor | [CLI Args Split Map](guides/cli-args-split-map.md) | M25 phased module-boundary plan to reduce `cli_args.rs` below 3000 LOC with API/test migration guardrails |
+| Runtime contributor | [Tools Runtime Split Map](guides/tools-split-map.md) | M25 phased module-boundary plan to reduce `tools.rs` below 3000 LOC with API/test migration guardrails |
+| Runtime contributor | [Benchmark Artifact Split Map](guides/benchmark-artifact-split-map.md) | M25 phased module-boundary plan to reduce `benchmark_artifact.rs` below 3000 LOC with API/test migration guardrails |
+| Runtime contributor | [GitHub Issues Runtime Split Map](guides/github-issues-runtime-split-map.md) | M25 phased module-boundary plan to reduce `github_issues_runtime.rs` below 3000 LOC with API/test migration guardrails |
+| Runtime contributor | [Channel Store Admin Split Map](guides/channel-store-admin-split-map.md) | M25 phased module-boundary plan to reduce `channel_store_admin.rs` below 2200 LOC with API/test migration guardrails |
 | Runtime contributor / doc maintainer | [Runbook Ownership Map](guides/runbook-ownership-map.md) | Post-consolidation runbook-to-crate ownership matrix and update policy |
 | Runtime operator / release manager | [Consolidated Runtime Rollback Drill](guides/consolidated-runtime-rollback-drill.md) | Trigger conditions, rollback simulation checklist, and required artifact capture flow |
 | Runtime operator / integration engineer | [MCP Client Operations Guide](guides/mcp-client-ops.md) | MCP client transport config (stdio + HTTP/SSE), OAuth PKCE token handling, inspect diagnostics, and live validation flow |
