+++ specs/2044/tasks.md
@@ -1,0 +1,13 @@
+# Tasks #2044
+
+Status: Completed
+Spec: specs/2044/spec.md
+Plan: specs/2044/plan.md
+
+## Ordered Tasks
+
+- T1: Complete split-map planning artifacts and contract validation (`#2066`).
+- T2: Extract operator-control summary/diff domain and enforce `<2200` guardrail
+  (`#2067`).
+- T3: Run targeted unit/functional/integration/regression evidence commands.
+- T4: Update parent spec status and close parent task with roll-up proof.
