+++ specs/2066/tasks.md
@@ -1,0 +1,12 @@
+# Tasks #2066
+
+Status: Completed
+Spec: specs/2066/spec.md
+Plan: specs/2066/plan.md
+
+## Ordered Tasks
+
+- T1: Add failing split-map contract checks and deterministic generation tests.
+- T2: Implement schema, generator, guide, and canonical report artifacts.
+- T3: Run shell + Python split-map validation suites (GREEN).
+- T4: Update documentation index/spec artifacts and post closure evidence.
