+++ specs/2067/tasks.md
@@ -1,0 +1,13 @@
+# Tasks #2067
+
+Status: Completed
+Spec: specs/2067/spec.md
+Plan: specs/2067/plan.md
+
+## Ordered Tasks
+
+- T1: Update/add failing guardrail assertions for M25.3.5 threshold and module
+  markers.
+- T2: Extract operator-control summary/diff workflow into helper module.
+- T3: Run targeted unit/functional/integration/regression verification.
+- T4: Record evidence and close `#2067` with parent roll-up in `#2044`.
